changed location of exported styled components (#1357)
* changed location of exported styled components

* changed location of exports
