+++ app/src/components/market/common/common_styled/index.tsx
@@ -1,11 +1,12 @@
 import styled, { css } from 'styled-components'
 
 import { getOutcomeColor } from '../../../../theme/utils'
-import { Button, ButtonContainer } from '../../../button'
+import { ButtonContainer } from '../../../button'
 
-export const LeftButton = styled(Button)`
-  margin-right: auto;
-`
+interface StatefulRadioButton {
+  selected?: boolean
+  disabled?: boolean
+}
 
 export const ButtonContainerFullWidth = styled(ButtonContainer)`
   margin-left: -${props => props.theme.cards.paddingHorizontal};
@@ -212,4 +213,65 @@ export const PercentWrapper = styled.label`
 export const CurrenciesWrapper = styled.div`
   padding: 0 0 20px 0;
   width: 100%;
+`
+export const CurationRow = styled.div`
+  border-bottom: ${props => props.theme.cards.border};
+  margin: 0 -25px;
+  padding: 20px 25px;
+  position: relative;
+`
+export const CurationSubRow = styled.div`
+  align-items: center;
+  display: flex;
+  flex-wrap: nowrap;
+  position: relative;
+`
+export const CurationLeftColumn = styled.div`
+  margin-right: 16px;
+`
+
+export const CurationCenterColumn = styled.div``
+
+export const CurationRightColumn = styled.div`
+  margin-left: auto;
+  text-align: right;
+  color: ${props => props.theme.colors.textColorDar};
+  font-weight: 500;
+`
+
+export const CurationLogoWrapper = styled.div`
+  padding: 11px;
+  border-radius: 50%;
+  border: 1px solid ${props => props.theme.colors.tertiary};
+  width: 48px;
+  height: 48px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`
+
+export const CurationRadioWrapper = styled.div<StatefulRadioButton>`
+  border-radius: 50%;
+  border: 1px solid ${props => props.theme.buttonPrimaryLine.borderColorDisabled};
+  cursor: ${props => (props.disabled ? 'initial' : 'pointer')};
+  width: 38px;
+  height: 38px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: ${props => props.selected && props.theme.colors.clickable};
+
+  &:hover {
+    border-color: ${props => !props.disabled && props.theme.colors.tertiary};
+  }
+`
+
+export const CurationOption = styled.div`
+  color: ${props => props.theme.colors.textColorDarker};
+  font-weight: 500;
+`
+
+export const CurationOptionDetails = styled.div`
+  color: ${props => props.theme.colors.textColorLighter};
+  font-weight: 400;
 `
