+++ app/src/components/market/sections/market_verify/market_verify.tsx
@@ -12,7 +12,7 @@ import { MarketMakerData, Status } from '../../../../util/types'
 import { Button, ButtonContainer } from '../../../button'
 import { ButtonType } from '../../../button/button_styling_types'
 import { InlineLoading } from '../../../loading'
-import { GenericError } from '../../common/common_styled'
+import { CurationRow, GenericError } from '../../common/common_styled'
 
 import { DxDaoCuration } from './option/dxdao_curation'
 import { KlerosCuration } from './option/kleros_curation'
@@ -30,80 +30,6 @@ const MarketVerification = styled.div`
   padding: 0 24px 0;
   border-top: ${({ theme }) => theme.borders.borderLineDisabled};
 `
-
-export const CurationRow = styled.div`
-  border-bottom: ${props => props.theme.cards.border};
-  margin: 0 -25px;
-  padding: 20px 25px;
-  position: relative;
-`
-
-export const CurationSubRow = styled.div`
-  align-items: center;
-  display: flex;
-  flex-wrap: no-wrap;
-  position: relative;
-`
-
-export const CurationLeftColumn = styled.div`
-  margin-right: 16px;
-`
-
-export const CurationCenterColumn = styled.div``
-
-export const CurationRightColumn = styled.div`
-  margin-left: auto;
-  text-align: right;
-  color: ${props => props.theme.colors.textColorDar};
-  font-weight: 500;
-`
-
-export const CurationRadioTick = styled.img<StatefulRadioButton>`
-  filter: ${props => (props.selected ? 'saturate(0) brightness(2)' : props.disabled ? 'saturate(0)' : '')};
-  cursor: ${props => (props.disabled ? 'initial' : 'pointer')};
-`
-
-export const CurationLogoWrapper = styled.div`
-  padding: 11px;
-  border-radius: 50%;
-  border: 1px solid ${props => props.theme.colors.tertiary};
-  width: 48px;
-  height: 48px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`
-
-export const CurationRadioWrapper = styled.div<StatefulRadioButton>`
-  border-radius: 50%;
-  border: ${props => (props.selected ? '' : `1px solid ${props.theme.buttonPrimaryLine.borderColorDisabled}`)};
-  cursor: ${props => (props.disabled ? 'initial' : 'pointer')};
-  width: 40px;
-  height: 40px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: ${props => props.selected && props.theme.colors.clickable};
-
-  &:hover {
-    border-color: ${props => !props.disabled && props.theme.colors.tertiary};
-  }
-`
-
-export const CurationOption = styled.div`
-  color: ${props => props.theme.colors.textColorDarker};
-  font-weight: 500;
-`
-
-export const CurationOptionDetails = styled.div`
-  color: ${props => props.theme.colors.textColorLighter};
-  font-weight: 400;
-`
-
-interface StatefulRadioButton {
-  selected?: boolean
-  disabled?: boolean
-}
 
 interface Props extends RouteComponentProps<any> {
   context: ConnectedWeb3Context
