+++ app/src/components/market/sections/market_verify/option/dxdao_curation.tsx
@@ -12,7 +12,7 @@ import {
   CurationRightColumn,
   CurationRow,
   CurationSubRow,
-} from '../market_verify'
+} from '../../../common/common_styled'
 
 const Bold = styled.b`
   font-weight: 500;
